+++ CHANGELOG.rst
@@ -1,5 +1,7 @@
 3.22:
 
+* `#459 <https://github.com/xitrum-framework/xitrum/issues/459>`_
+  Optimize route collecting: Ignore more packages that obviously don't contain routes
 * `#495 <https://github.com/xitrum-framework/xitrum/issues/495>`_
   Update Netty from 4.0.24 to 4.0.25